Refactor vocabcompiler to reduce dependence on core code for plugins
